combined all in class GitHubService and SlackService change GitSlackUserController
