+++ src/main/java/com/community/tools/controller/GitSlackUsersController.java
@@ -2,8 +2,8 @@ package com.community.tools.controller;
 
 import static org.springframework.http.ResponseEntity.ok;
 
-import com.community.tools.service.github.GitHubUsersService;
-import com.community.tools.service.slack.SlackUsersService;
+import com.community.tools.service.github.GitHubService;
+import com.community.tools.service.slack.SlackService;
 import com.github.seratch.jslack.api.model.User;
 import com.github.seratch.jslack.api.model.User.Profile;
 import java.util.List;
@@ -15,36 +15,35 @@ import org.kohsuke.github.GHUser;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
 @RestController
+@RequestMapping("app")
 public class GitSlackUsersController {
 
-  private final SlackUsersService usersService;
-  private final GitHubUsersService gitUsersService;
+  private final SlackService usersService;
+  private final GitHubService gitService;
 
-  @GetMapping(value = "/users", produces = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<List<String>> getAllUsers(@RequestParam(name = "app") String app) {
-    if (app.equals("git")) {
-      Set<GHUser> gitHubAllUsers = gitUsersService.getGitHubAllUsers();
+  @GetMapping(value = "/git", produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<List<String>> getGitHubAllUsers() {
+    Set<GHUser> gitHubAllUsers = gitService.getGitHubAllUsers();
 
-      List<String> listGitUsersLogin = gitHubAllUsers.stream().map(GHPerson::getLogin)
-          .collect(Collectors.toList());
+    List<String> listGitUsersLogin = gitHubAllUsers.stream().map(GHPerson::getLogin)
+        .collect(Collectors.toList());
 
-      return ok().body(listGitUsersLogin);
-    }
+    return ok().body(listGitUsersLogin);
+  }
 
-    if (app.equals("slack")) {
-      Set<User> allSlackUsers = usersService.getAllUsers();
+  @GetMapping(value = "/slack", produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<List<String>> getSlackAllUsers() {
+    Set<User> allSlackUsers = usersService.getAllUsers();
 
-      List<String> listSlackUsersName = allSlackUsers.stream()
-          .map(User::getProfile)
-          .map(Profile::getDisplayName).collect(Collectors.toList());
+    List<String> listSlackUsersName = allSlackUsers.stream()
+        .map(User::getProfile)
+        .map(Profile::getDisplayName).collect(Collectors.toList());
 
-      return ok().body(listSlackUsersName);
-    }
-    throw new RuntimeException("you enter wrong app");
+    return ok().body(listSlackUsersName);
   }
 }
